refactor mount-s3 script to check for command existence before installation

+++ .platform/hooks/prebuild/20-mount-s3.sh
@@ -7,14 +7,21 @@ if [[ -z "$CONFIG_BUCKET_NAME" ]]; then
   exit 1
 fi
 
-MP_RPM=$(mktemp --suffix=.rpm)
-curl https://s3.amazonaws.com/mountpoint-s3-release/latest/arm64/mount-s3.rpm > $MP_RPM
-yum install -y $MP_RPM
-rm $MP_RPM
+if [ ! command -v mount-s3 &> /dev/null ]; then
+  MP_RPM=$(mktemp --suffix=.rpm)
+  curl https://s3.amazonaws.com/mountpoint-s3-release/latest/arm64/mount-s3.rpm > $MP_RPM
+  yum install -y $MP_RPM
+  rm $MP_RPM
+fi
 
 MNT_PATH=/mnt/s3/$CONFIG_BUCKET_NAME
-echo "s3://$CONFIG_BUCKET_NAME/ ${MNT_PATH} mount-s3 _netdev,nosuid,nodev,rw,allow-other,nofail" >> /etc/fstab
-mkdir -p $MNT_PATH
+echo "MNT_PATH=$MNT_PATH" >> .env
 
-systemctl daemon-reload
-mount -a
+if ! grep -q "s3://$CONFIG_BUCKET_NAME/" /etc/fstab; then
+  echo "s3://$CONFIG_BUCKET_NAME/ ${MNT_PATH} mount-s3 _netdev,nosuid,nodev,rw,allow-other,nofail" >> /etc/fstab
+  mkdir -p $MNT_PATH
+  systemctl daemon-reload
+  mount -a
+fi
+
+
